dictionary/requirements.txt
django-markitup